+++ public/index.html
@@ -8,7 +8,7 @@
     .##.....##.##.....##.##....##.##...##..##....##..##.....##....##....##....##..........##
     .##.....##.##.....##..######..##....##.##.....##..#######......######...######..#######.
 
-    ©2019 HackRU Team at Rutgers University
+    Copyright © 2019 HackRU Team at Rutgers University
     Frontend: HackRU Frontend Team
     Backend: HackRU Research and Development Team (RnD)
 -->
@@ -22,7 +22,8 @@
     <meta property="og:image" content="./assets/fb-banner-upper.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="description" content="Sign up for HackRU Spring 2019, taking place from March 9th to 10th.">
-    <link rel="shortcut icon" href="%PUBLIC_URL%/assets/fav.png">
+    <link rel="icon" href="%PUBLIC_URL%/assets/icons/hru-logo-circle-green.png">
+    <link rel="shortcut icon" href="%PUBLIC_URL%/assets/icons/hru-logo-circle-green.png">
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
     <link rel="stylesheet" href="%PUBLIC_URL%/index.css">
     <title id="title">HackRU</title>
