+++ public/manifest.json
@@ -1,15 +1,15 @@
 {
-  "short_name": "HackRU",
-  "name": "HackRU Spring 2019",
-  "icons": [
-    {
-      "src": "favicon.ico",
-      "sizes": "64x64 32x32 24x24 16x16",
-      "type": "image/x-icon"
-    }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
+    "short_name": "HackRU",
+    "name": "HackRU Spring 2019",
+    "icons": [
+        {
+            "src": "./assets/icons/hru-logo-circle-green.png",
+            "sizes": "64x64 32x32 24x24 16x16",
+            "type": "image/png"
+        }
+    ],
+    "start_url": ".",
+    "display": "standalone",
+    "theme_color": "#000000",
+    "background_color": "#ffffff"
 }
