+++ src/components/Dashboard/Resources/majors.json
