+++ src/components/Dashboard/Section.jsx
@@ -2,9 +2,10 @@ import React, { Component } from 'react';
 import { Collapse, ListGroup, ListGroupItem, ListGroupItemHeading, ListGroupItemText } from 'reactstrap';
 import { Icon } from "react-fa";
 import { theme } from "../../Defaults";
-
-// A collapsable section on the dashboard
-// It takes four props, 'children', 'isOpen' (if it is initially expanded), 'title', and 'subtitle'
+/**
+ * A collapsable section on the dashboard
+ * It takes four props, 'children', 'isOpen' (if it is initially expanded), 'title', and 'subtitle'
+ */
 class Section extends Component {
     constructor(props) {
         super(props)
@@ -12,42 +13,30 @@ class Section extends Component {
             isOpen: props.isOpen
         }
     }
-
     render() {
         const { children, title, subtitle, isOpen, ...rest } = this.props;
         return (
-            <div
-                style={{ width: "100%", textAlign: "left", marginBottom: 25 }}
-                {...rest}
-            >
+            <div style={{ width: "100%", textAlign: "left", marginBottom: 25 }} {...rest} >
                 <ListGroup>
-                    <ListGroupItem
-                        action 
-                        tag="button"
-                        href="#"
-                        style={{ background: theme.primary[0] + "2F", color: theme.primary[1] + "FF", borderRadius: 0 }}
-                        onClick={(e) => 
-                            this.setState({ isOpen: !this.state.isOpen })
-                        }
-                    >
+                    <ListGroupItem action tag="button" href="#" style={{ background: theme.primary[0] + "2F", color: theme.primary[1] + "FF", borderRadius: 0 }} onClick={(e) => 
+                        this.setState({ isOpen: !this.state.isOpen })
+                    }>
                         <ListGroupItemHeading>
-                            { title }
+                            {title}
                         </ListGroupItemHeading>
                         <ListGroupItemText>
-                            { subtitle }
+                            {subtitle}
                         </ListGroupItemText>
                         <Icon style={{ position: "absolute", right: 25, top: 25 }} name={(this.state.isOpen) ? ("chevron-up") : ("chevron-down")} />
                     </ListGroupItem>
                     <Collapse isOpen={this.state.isOpen}>
                         <ListGroupItem style={{ background: theme.primary[1] + "1F", borderRadius: 0 }}>
-                            { children }
+                            {children}
                         </ListGroupItem>
                     </Collapse>
                 </ListGroup>
             </div>
         )
     }
-
 }
-
 export default Section
